feat: isolate app, models and routing
